Removed DatafeedLoop ABC and reorganized BDK config loader modules
