+++ examples/core/config/config_main.py
@@ -1,4 +1,4 @@
-from symphony.bdk.core.config.bdk_config_loader import BdkConfigLoader
+from symphony.bdk.core.config.loader import BdkConfigLoader
 
 
 class ConfigMain:
