+++ symphony/bdk/core/config/loader.py
@@ -1,6 +1,8 @@
+import json
 from pathlib import Path
 
-from symphony.bdk.core.config.bdk_config_parser import BdkConfigParser
+import yaml
+
 from symphony.bdk.core.config.exception import BdkConfigException
 from symphony.bdk.core.config.model.bdk_config import BdkConfig
 
@@ -49,3 +51,30 @@ class BdkConfigLoader:
         """
         config_path = (Path.home() / ".symphony" / relative_path).resolve()
         return cls.load_from_file(str(config_path))
+
+
+class BdkConfigParser:
+    """ Config Parser class
+
+    Provide methods to Deserialize a configuration content
+    as a ``str`` in a JSON or YAML format to a BdkConfig object.
+    """
+
+    @classmethod
+    def parse(cls, config_content: str):
+        """
+
+        :param config_content:
+        :return:
+        """
+        try:
+            return json.loads(config_content)
+        except json.JSONDecodeError:
+            pass
+
+        try:
+            return yaml.safe_load(config_content)
+        except yaml.YAMLError:
+            pass
+
+        raise BdkConfigException("Config file is neither in JSON nor in YAML format.")
