+++ tests/core/client/api_client_factory_test.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import pytest
 
 from symphony.bdk.core.client.api_client_factory import ApiClientFactory
-from symphony.bdk.core.config.bdk_config_loader import BdkConfigLoader
+from symphony.bdk.core.config.loader import BdkConfigLoader
 from tests.utils.resource_utils import get_config_resource_filepath
